Move cfunc_from_ea
